# perf(serde_v8): drop need for RC in serializer